Infrastructure - add repositories #43

+++ src/MonitorPet.Application/Model/PesoHistorico/WeightHistoryModel.cs
@@ -3,6 +3,6 @@
 public class WeightHistoryModel
 {
     public long Id { get; set; }
-    public long Weight { get; set; }
+    public double Weight { get; set; }
     public DateTimeOffset RegisteredAt { get; set; }
 }
